add support for name and avatar support
adds support for setting TOKEN_APP_NAME and TOKEN_APP_AVATAR to set
name and avatar for the bot.

If TOKEN_APP_AVATAR is a URL (starting with http or https) it will
post it in the normal user update. If it is a file, it will use the
file upload endpoint to set the avatar to the given file (if it exists
and is a supported mimetype).